Collision resolution function table
