+++ include/physics/collider.h
@@ -18,21 +18,3 @@ struct Collider{
   ColliderType type;
   union ColliderData data;
 };
-
-// struct Collider {
-//   COLLIDER_TYPE type;
-//   // Thanks Low Level
-//   union {
-//     struct AABB aabb;
-//     struct Plane plane;
-//   } data;
-//   vec3 position;
-//   // Spheres wouldn't care about rotation... but whatever
-//   vec3 rotation;
-//   vec3 velocity;
-// };
-
-struct PlaneCollider {
-  vec3 normal;
-  float distance;
-};
